# docs/openapi-public.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Community Feed API",
    "description": "Corgi Commons is a Bluesky feed with inspectable, community-shaped ranking. Approved pilot participants can vote on ranking signals, topic priorities, and content rules; reviewed policy changes are applied before rescoring. Built on AT Protocol.\n\n## Authentication\n- **Governance endpoints** require a session cookie or bearer token from `POST /api/governance/auth/login`.\n- **Admin endpoints** additionally require the caller's DID to be in the `BOT_ADMIN_DIDS` allowlist.\n- **Feed endpoints** are public (called by the Bluesky app). Viewing the feed and using the shadow demo do not require governance access.\n- **Transparency endpoints** are public and unauthenticated.",
    "version": "1.2.0",
    "contact": {
      "name": "Community Feed",
      "url": "https://github.com/andrewnordstrom-eng/bluesky-community-feed"
    },
    "license": {
      "name": "Apache-2.0"
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "governance_session",
        "description": "Session cookie from POST /api/governance/auth/login"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token from POST /api/governance/auth/login (for CLI/MCP)"
      }
    },
...
  },
  "paths": {
...
    "/api/transparency/post/{uri}": {
      "get": {
        "summary": "Explain post ranking",
        "tags": [
          "Transparency"
        ],
        "description": "Returns a full breakdown of why a post is ranked where it is: all 5 component scores (raw, weight, weighted), an exact published position when the post is in the current materialized snapshot, or a legacy score-run rank otherwise, and a counterfactual showing what rank it would have with pure engagement sorting.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "uri",
            "required": true,
            "description": "AT-URI of the post (URL-encoded)"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "post_uri": {
                      "type": "string"
                    },
                    "epoch_id": {
                      "type": "integer"
                    },
                    "epoch_description": {
                      "type": "string",
                      "nullable": true
                    },
                    "total_score": {
                      "type": "number"
                    },
                    "rank": {
                      "type": "integer"
                    },
                    "rank_scope": {
                      "type": "string",
                      "enum": [
                        "published_snapshot",
                        "score_run"
                      ]
                    },
                    "published_position": {
                      "type": "integer",
                      "nullable": true,
                      "description": "Position in the published top-50 transparency presentation snapshot."
                    },
                    "publication_snapshot_id": {
                      "type": "string",
                      "nullable": true
                    },
                    "components": {
                      "type": "object",
                      "description": "All 5 scoring components with raw, weight, and weighted values",
                      "additionalProperties": {
                        "type": "object",
                        "properties": {
                          "raw_score": {
                            "type": "number"
                          },
                          "weight": {
                            "type": "number"
                          },
                          "weighted": {
                            "type": "number"
                          },
                          "topicBreakdown": {
                            "type": "object",
                            "additionalProperties": {
                              "type": "object",
                              "properties": {
                                "postScore": {
                                  "type": "number"
                                },
                                "communityWeight": {
                                  "type": "number"
                                },
                                "contribution": {
                                  "type": "number"
                                }
                              },
                              "required": [
                                "postScore",
                                "communityWeight",
                                "contribution"
                              ]
                            }
                          }
                        }
                      }
                    },
                    "governance_weights": {
                      "type": "object",
                      "description": "Per-component weight vector from the current governance epoch.",
                      "additionalProperties": {
                        "type": "number"
                      }
                    },
                    "counterfactual": {
                      "type": "object",
                      "properties": {
                        "pure_engagement_rank": {
                          "type": "integer"
                        },
                        "community_governed_rank": {
                          "type": "integer"
                        },
                        "difference": {
                          "type": "integer",
                          "description": "Positive = governance boosted this post"
                        }
                      }
                    },
                    "scored_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "classification_method": {
                      "type": "string",
                      "enum": [
                        "keyword",
                        "embedding"
                      ]
                    }
                  },
                  "required": [
                    "post_uri",
                    "epoch_id",
                    "total_score",
                    "rank",
                    "rank_scope",
                    "published_position",
                    "publication_snapshot_id",
                    "components"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error code (e.g. \"ValidationError\", \"NotFound\")"
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "correlationId": {
                      "type": "string",
                      "description": "Request correlation ID for debugging"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error code (e.g. \"ValidationError\", \"NotFound\")"
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "correlationId": {
                      "type": "string",
                      "description": "Request correlation ID for debugging"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error code (e.g. \"ValidationError\", \"NotFound\")"
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "correlationId": {
                      "type": "string",
                      "description": "Request correlation ID for debugging"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          },
          "503": {
            "description": "TransparencySnapshotIntegrityFailure when a published receipt artifact fails validation, or TransparencySnapshotUnavailable when current public eligibility cannot be established.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error code (e.g. \"ValidationError\", \"NotFound\")"
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "correlationId": {
                      "type": "string",
                      "description": "Request correlation ID for debugging"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "description": "TransparencySnapshotIntegrityFailure when a published receipt artifact fails validation, or TransparencySnapshotUnavailable when current public eligibility cannot be established."
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://feed.yourdomain.com",
      "description": "Production"
    }
  ],
...
}